Search GO terms for a matching pattern

This function searches the the Gene Ontology terms for a specific pattern. If run in interactive mode and if the pattern matches more then one term, the user is presented in a menu to choose one term.

searchGO(pattern, namespace = c("CC", "BP", "MF"), ignore.case = TRUE,
  interactive = TRUE, graphics = getOption("menu.graphics"), ...)

Arguments

pattern

A character of length 1 (subsequent elements are ignored with a warning) to search for in GO.

namespace

One or multiple GO namespaces to search for pattern. Must match "CC", "MF" and "BP".

ignore.case

Should case be ignored. Default is TRUE.

interactive

If TRUE (default) and several GO terms matched pattern, a menu is provided to the user to select one term.

graphics

Should a graphical menu be used. Defaults to getOption("menu.graphics").

...

Additional parameters, except ignore.case (see above), passed to grep.

Value

A character with the identifiers of matched GO terms of invisibly NULL if no matches were found.

Details

This function searches the identifier, term and synonym fields of all Gene Ontology or only a subset of the possible namespaces. The pattern is cases insensitive, unless otherwise stated, and more pattern matching can be further parametrised by passing additional arguments to the underlying grep function.

If more than one match is found and the function is used interactively, the user is presented with a menu to select a single term. If the users exits the menu without any selection (by choosing 0 or cancel), all matching identifiers are returned, as if the search was performed in non-interactive mode.

If not match is found, NULL is returned invisibly with a message.

Examples

i <- interactive()
searchGO("Golgi", interactive = i)

#> Loading required package: GO.db

#> Loading required package: AnnotationDbi

#> Loading required package: stats4

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, cbind, colMeans, colnames,
#>     colSums, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, lengths, Map, mapply, match,
#>     mget, order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rowMeans, rownames, rowSums, sapply, setdiff, sort,
#>     table, tapply, union, unique, unsplit, which, which.max, which.min

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: IRanges

#> Loading required package: S4Vectors

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> 
#> Attaching package: ‘AnnotationDbi’

#> The following object is masked from ‘package:yagop’:
#> 
#>     GOTerms

#> Select a GO term 
#> 
#>   1: GO:0000042 - protein targeting to Golgi (BP)
#>   2: GO:0000137 - Golgi cis cisterna (CC)
#>   3: GO:0000138 - Golgi trans cisterna (CC)
#>   4: GO:0000139 - Golgi membrane (CC)
#>   5: GO:0000301 - retrograde transport, vesicle recycling within Golgi (BP)
#>   6: GO:0005793 - endoplasmic reticulum-Golgi intermediate compartment (CC)
#>   7: GO:0005794 - Golgi apparatus (CC)
#>   8: GO:0005795 - Golgi stack (CC)
#>   9: GO:0005796 - Golgi lumen (CC)
#>  10: GO:0005797 - Golgi medial cisterna (CC)
#>  11: GO:0005798 - Golgi-associated vesicle (CC)
#>  12: GO:0005801 - cis-Golgi network (CC)
#>  13: GO:0005802 - trans-Golgi network (CC)
#>  14: GO:0006888 - ER to Golgi vesicle-mediated transport (BP)
#>  15: GO:0006890 - retrograde vesicle-mediated transport, Golgi to ER (BP)
#>  16: GO:0006891 - intra-Golgi vesicle-mediated transport (BP)
#>  17: GO:0006892 - post-Golgi vesicle-mediated transport (BP)
#>  18: GO:0006893 - Golgi to plasma membrane transport (BP)
#>  19: GO:0006895 - Golgi to endosome transport (BP)
#>  20: GO:0006896 - Golgi to vacuole transport (BP)
#>  21: GO:0007030 - Golgi organization (BP)
#>  22: GO:0010785 - clathrin coating of Golgi vesicle, plasma membrane to endosome targeting (BP)
#>  23: GO:0010786 - clathrin coating of Golgi vesicle, trans-Golgi to endosome targeting (BP)
#>  24: GO:0010787 - COPI coating of Golgi vesicle, inter-Golgi cisterna (BP)
#>  25: GO:0010788 - COPI coating of Golgi vesicle, cis-Golgi to rough ER (BP)
#>  26: GO:0012507 - ER to Golgi transport vesicle membrane (CC)
#>  27: GO:0012508 - Golgi to ER transport vesicle membrane (CC)
#>  28: GO:0012509 - inter-Golgi transport vesicle membrane (CC)
#>  29: GO:0012510 - trans-Golgi network transport vesicle membrane (CC)
#>  30: GO:0017119 - Golgi transport complex (CC)
#>  31: GO:0021701 - cerebellar Golgi cell differentiation (BP)
#>  32: GO:0021926 - Golgi cell precursor proliferation (BP)
#>  33: GO:0030130 - clathrin coat of trans-Golgi network vesicle (CC)
#>  34: GO:0030134 - ER to Golgi transport vesicle (CC)
#>  35: GO:0030140 - trans-Golgi network transport vesicle (CC)
#>  36: GO:0030142 - Golgi to ER transport vesicle (CC)
#>  37: GO:0030143 - inter-Golgi transport vesicle (CC)
#>  38: GO:0030173 - integral component of Golgi membrane (CC)
#>  39: GO:0030660 - Golgi-associated vesicle membrane (CC)
#>  40: GO:0031228 - intrinsic component of Golgi membrane (CC)
#>  41: GO:0031985 - Golgi cisterna (CC)
#>  42: GO:0032468 - Golgi calcium ion homeostasis (BP)
#>  43: GO:0032472 - Golgi calcium ion transport (BP)
#>  44: GO:0032580 - Golgi cisterna membrane (CC)
#>  45: GO:0032588 - trans-Golgi network membrane (CC)
#>  46: GO:0033106 - cis-Golgi network membrane (CC)
#>  47: GO:0033116 - endoplasmic reticulum-Golgi intermediate compartment membrane (CC)
#>  48: GO:0033578 - protein glycosylation in Golgi (BP)
#>  49: GO:0033581 - protein galactosylation in Golgi (BP)
#>  50: GO:0034067 - protein localization to Golgi apparatus (BP)
#>  51: GO:0034469 - Golgi stack lumen (CC)
#>  52: GO:0034498 - early endosome to Golgi transport (BP)
#>  53: GO:0034499 - late endosome to Golgi transport (BP)
#>  54: GO:0035526 - retrograde transport, plasma membrane to Golgi (BP)
#>  55: GO:0035621 - ER to Golgi ceramide transport (BP)
#>  56: GO:0036085 - GDP-fucose import into Golgi lumen (BP)
#>  57: GO:0036504 - Golgi membrane fusion (BP)
#>  58: GO:0042147 - retrograde transport, endosome to Golgi (BP)
#>  59: GO:0042996 - regulation of Golgi to plasma membrane protein transport (BP)
#>  60: GO:0042997 - negative regulation of Golgi to plasma membrane protein transport (BP)
#>  61: GO:0042998 - positive regulation of Golgi to plasma membrane protein transport (BP)
#>  62: GO:0042999 - regulation of Golgi to plasma membrane CFTR protein transport (BP)
#>  63: GO:0043000 - Golgi to plasma membrane CFTR protein transport (BP)
#>  64: GO:0043001 - Golgi to plasma membrane protein transport (BP)
#>  65: GO:0043002 - negative regulation of Golgi to plasma membrane CFTR protein transport (BP)
#>  66: GO:0043003 - positive regulation of Golgi to plasma membrane CFTR protein transport (BP)
#>  67: GO:0044172 - host cell endoplasmic reticulum-Golgi intermediate compartment (CC)
#>  68: GO:0044173 - host cell endoplasmic reticulum-Golgi intermediate compartment membrane (CC)
#>  69: GO:0044177 - host cell Golgi apparatus (CC)
#>  70: GO:0044178 - host cell Golgi membrane (CC)
#>  71: GO:0044431 - Golgi apparatus part (CC)
#>  72: GO:0044795 - trans-Golgi network to recycling endosome transport (BP)
#>  73: GO:0045018 - retrograde transport, vacuole to Golgi (BP)
#>  74: GO:0045053 - protein retention in Golgi apparatus (BP)
#>  75: GO:0046760 - viral budding from Golgi membrane (BP)
#>  76: GO:0048193 - Golgi vesicle transport (BP)
#>  77: GO:0048194 - Golgi vesicle budding (BP)
#>  78: GO:0048195 - Golgi membrane priming complex assembly (BP)
#>  79: GO:0048197 - Golgi membrane coat protein complex assembly (BP)
#>  80: GO:0048198 - Golgi vesicle bud deformation and release (BP)
#>  81: GO:0048199 - vesicle targeting, to, from or within Golgi (BP)
#>  82: GO:0048200 - Golgi transport vesicle coating (BP)
#>  83: GO:0048202 - clathrin coating of Golgi vesicle (BP)
#>  84: GO:0048203 - vesicle targeting, trans-Golgi to endosome (BP)
#>  85: GO:0048204 - vesicle targeting, inter-Golgi cisterna (BP)
#>  86: GO:0048205 - COPI coating of Golgi vesicle (BP)
#>  87: GO:0048206 - vesicle targeting, cis-Golgi to rough ER (BP)
#>  88: GO:0048207 - vesicle targeting, rough ER to cis-Golgi (BP)
#>  89: GO:0048209 - regulation of vesicle targeting, to, from or within Golgi (BP)
#>  90: GO:0048210 - Golgi vesicle fusion to target membrane (BP)
#>  91: GO:0048211 - Golgi vesicle docking (BP)
#>  92: GO:0048212 - Golgi vesicle uncoating (BP)
#>  93: GO:0048213 - Golgi vesicle prefusion complex stabilization (BP)
#>  94: GO:0048214 - regulation of Golgi vesicle fusion to target membrane (BP)
#>  95: GO:0048215 - positive regulation of Golgi vesicle fusion to target membrane (BP)
#>  96: GO:0048216 - negative regulation of Golgi vesicle fusion to target membrane (BP)
#>  97: GO:0048219 - inter-Golgi cisterna vesicle-mediated transport (BP)
#>  98: GO:0048280 - vesicle fusion with Golgi apparatus (BP)
#>  99: GO:0048313 - Golgi inheritance (BP)
#> 100: GO:0051645 - Golgi localization (BP)
#> 101: GO:0051683 - establishment of Golgi localization (BP)
#> 102: GO:0051684 - maintenance of Golgi location (BP)
#> 103: GO:0055107 - Golgi to secretory granule transport (BP)
#> 104: GO:0055108 - Golgi to transport vesicle transport (BP)
#> 105: GO:0060628 - regulation of ER to Golgi vesicle-mediated transport (BP)
#> 106: GO:0061454 - Golgi calcium ion export (BP)
#> 107: GO:0070319 - Golgi to plasma membrane transport vesicle (CC)
#> 108: GO:0070931 - Golgi-associated vesicle lumen (CC)
#> 109: GO:0071955 - recycling endosome to Golgi transport (BP)
#> 110: GO:0072600 - establishment of protein localization to Golgi (BP)
#> 111: GO:0090113 - regulation of ER to Golgi vesicle-mediated transport by GTP hydrolysis (BP)
#> 112: GO:0090160 - Golgi to lysosome transport (BP)
#> 113: GO:0090161 - Golgi ribbon formation (BP)
#> 114: GO:0090164 - asymmetric Golgi ribbon formation (BP)
#> 115: GO:0090165 - regulation of secretion by asymmetric Golgi ribbon formation (BP)
#> 116: GO:0090166 - Golgi disassembly (BP)
#> 117: GO:0090167 - Golgi distribution to daughter cells (BP)
#> 118: GO:0090168 - Golgi reassembly (BP)
#> 119: GO:0090170 - regulation of Golgi inheritance (BP)
#> 120: GO:0090283 - regulation of protein glycosylation in Golgi (BP)
#> 121: GO:0090284 - positive regulation of protein glycosylation in Golgi (BP)
#> 122: GO:0090285 - negative regulation of protein glycosylation in Golgi (BP)
#> 123: GO:0090498 - extrinsic component of Golgi membrane (CC)
#> 124: GO:0097111 - endoplasmic reticulum-Golgi intermediate compartment organization (BP)
#> 125: GO:0097624 - UDP-galactose transmembrane import into Golgi lumen (BP)
#> 126: GO:0097712 - vesicle targeting, trans-Golgi to periciliary membrane compartment (BP)
#> 127: GO:0098540 - lumenal side of trans-Golgi network transport vesicle membrane (CC)
#> 128: GO:0098541 - cytoplasmic side of trans-Golgi network transport vesicle membrane (CC)
#> 129: GO:0098547 - lumenal side of Golgi membrane (CC)
#> 130: GO:0098548 - cytoplasmic side of Golgi membrane (CC)
#> 131: GO:0098564 - trans-Golgi network transport vesicle lumen (CC)
#> 132: GO:0098629 - trans-Golgi network membrane organization (BP)
#> 133: GO:0098791 - Golgi subcompartment (CC)
#> 134: GO:0099041 - vesicle tethering to Golgi (BP)
#> 135: GO:1901472 - regulation of Golgi calcium ion export (BP)
#> 136: GO:1902953 - positive regulation of ER to Golgi vesicle-mediated transport (BP)
#> 137: GO:1903292 - protein localization to Golgi membrane (BP)
#> 138: GO:1903358 - regulation of Golgi organization (BP)
#> 139: GO:1904257 - zinc II ion import across Golgi membrane (BP)
#> 140: GO:1904381 - Golgi apparatus mannose trimming (BP)
#> 141: GO:1905279 - regulation of retrograde transport, endosome to Golgi (BP)
#> 142: GO:1905280 - negative regulation of retrograde transport, endosome to Golgi (BP)
#> 143: GO:1905281 - positive regulation of retrograde transport, endosome to Golgi (BP)
#> 144: GO:1990536 - phosphoenolpyruvate transmembrane import into Golgi lumen (BP)
#> 145: GO:1990668 - vesicle fusion with endoplasmic reticulum-Golgi intermediate compartment (ERGIC) membrane (BP)
#> 146: GO:1990669 - endoplasmic reticulum-Golgi intermediate compartment (ERGIC) derived vesicle fusion with endoplasmic reticulum membrane (BP)
#> 147: GO:1990670 - vesicle fusion with Golgi cis cisterna membrane (BP)
#> 148: GO:1990671 - vesicle fusion with Golgi medial cisterna membrane (BP)
#> 149: GO:1990672 - medial-Golgi-derived vesicle fusion with Golgi trans cisterna membrane (BP)
#> 150: GO:1990673 - intrinsic component of endoplasmic reticulum-Golgi intermediate compartment (ERGIC) membrane (CC)
#> 151: GO:1990674 - Golgi cis cisterna membrane (CC)
#> 152: GO:1990675 - Golgi medial cisterna membrane (CC)
#> 153: GO:1990676 - Golgi trans cisterna membrane (CC)
#> 154: GO:1990687 - endoplasmic reticulum-derived vesicle fusion with endoplasmic reticulum-Golgi intermediate compartment (ERGIC) membrane (BP)
#> 155: GO:1990688 - Golgi vesicle fusion with endoplasmic reticulum-Golgi intermediate compartment (ERGIC) membrane (BP)
#> 156: GO:1990689 - endoplasmic reticulum-Golgi intermediate compartment (ERGIC) derived vesicle fusion with Golgi cis cisterna membrane (BP)
#> 157: GO:1990690 - Golgi medial cisterna-derived vesicle fusion with Golgi cis cisterna membrane (BP)
#> 158: GO:1990691 - cis-Golgi-derived vesicle fusion with Golgi medial cisterna membrane (BP)
#> 159: GO:1990692 - trans-Golgi-derived vesicle fusion with Golgi medial cisterna membrane (BP)
#> 160: GO:1990693 - intrinsic component of Golgi cis cisterna membrane (CC)
#> 161: GO:1990694 - intrinsic component of Golgi medial cisterna membrane (CC)
#> 162: GO:1990695 - intrinsic component of Golgi trans cisterna membrane (CC)
#> 163: GO:1990701 - integral component of endoplasmic reticulum-Golgi intermediate compartment (ERGIC) membrane (CC)
#> 164: GO:1990702 - integral component of Golgi cis cisterna membrane (CC)
#> 165: GO:1990703 - integral component of Golgi medial cisterna membrane (CC)
#> 166: GO:1990704 - integral component of Golgi trans cisterna membrane (CC)
#> 167: GO:2000156 - regulation of retrograde vesicle-mediated transport, Golgi to ER (BP)
#> 168: GO:0000938 - GARP complex (CC)
#> 169: GO:0004572 - mannosyl-oligosaccharide 1,3-1,6-alpha-mannosidase activity (MF)
#> 170: GO:0030133 - transport vesicle (CC)
#> 171: GO:0043529 - GET complex (CC)
#> 172: GO:0046745 - viral capsid secondary envelopment (BP)
#> 
#> Enter an item from the menu, or 0 to exit
#> Enter an item from the menu, or 0 to exit

#> [1] "GO:0000042"

searchGO("Golgi apparatus", interactive = i)

#> Select a GO term 
#> 
#>  1: GO:0005794 - Golgi apparatus (CC)
#>  2: GO:0034067 - protein localization to Golgi apparatus (BP)
#>  3: GO:0044177 - host cell Golgi apparatus (CC)
#>  4: GO:0044431 - Golgi apparatus part (CC)
#>  5: GO:0045053 - protein retention in Golgi apparatus (BP)
#>  6: GO:0048280 - vesicle fusion with Golgi apparatus (BP)
#>  7: GO:1904381 - Golgi apparatus mannose trimming (BP)
#>  8: GO:0051645 - Golgi localization (BP)
#>  9: GO:0051683 - establishment of Golgi localization (BP)
#> 10: GO:0051684 - maintenance of Golgi location (BP)
#> 11: GO:0072600 - establishment of protein localization to Golgi (BP)
#> 
#> Enter an item from the menu, or 0 to exit

#> [1] "GO:0005794"

searchGO("Golgi apparatus", namespace = "CC", interactive = i)

#> Select a GO term 
#> 
#> 1: GO:0005794 - Golgi apparatus (CC)
#> 2: GO:0044177 - host cell Golgi apparatus (CC)
#> 3: GO:0044431 - Golgi apparatus part (CC)
#> 
#> Enter an item from the menu, or 0 to exit

#> [1] "GO:0005794"

searchGO("Golgi apparatus", namespace = "BP", interactive = i)

#> Select a GO term 
#> 
#> 1: GO:0034067 - protein localization to Golgi apparatus (BP)
#> 2: GO:0045053 - protein retention in Golgi apparatus (BP)
#> 3: GO:0048280 - vesicle fusion with Golgi apparatus (BP)
#> 4: GO:1904381 - Golgi apparatus mannose trimming (BP)
#> 5: GO:0051645 - Golgi localization (BP)
#> 6: GO:0051683 - establishment of Golgi localization (BP)
#> 7: GO:0051684 - maintenance of Golgi location (BP)
#> 8: GO:0072600 - establishment of protein localization to Golgi (BP)
#> 
#> Enter an item from the menu, or 0 to exit

#> [1] "GO:0034067"

searchGO("Golgi apparatus", namespace = "MF", interactive = i)

#> No matching GO entry found.